pyproject.toml
[tool.poetry]
name = "drf-auth-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
djangorestframework-jwt = "^1.11.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.1"
djangorestframework-simplejwt = "4.6"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
